requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
jsonschema>=4.20.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0

